+++ CHANGELOG.md
@@ -7,13 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `RECEIPT_ID` option to external id generation for forms
+
 ## [2.0.2] - 2023-03-13
 
 ### Added
 
 - `formApprovalFlowInstanceId` to `Approvals.searchFormSubmissionAdministrationApprovals()`
 - `requiredAll` to checkbox form element
-- `RECEIPT_ID` option to external id generation for forms
 
 ### Dependencies